detach all update_model to learner class for all agents
