+++ rl_algorithms/ddpg/learner.py
@@ -1,12 +1,12 @@
 import argparse
-from typing import Tuple, Union
+from typing import Tuple
 
 import torch
 import torch.nn.functional as F
 from torch.nn.utils import clip_grad_norm_
 import torch.optim as optim
 
-from rl_algorithms.common.abstract.learner import Learner, TensorTuple
+from rl_algorithms.common.abstract.learner import Learner
 import rl_algorithms.common.helper_functions as common_utils
 from rl_algorithms.common.networks.brain import Brain
 from rl_algorithms.utils.config import ConfigDict
@@ -18,30 +18,81 @@ class DDPGLearner(Learner):
     Attributes:
         args (argparse.Namespace): arguments including hyperparameters and training settings
         hyper_params (ConfigDict): hyper-parameters
+        optim_cfg (ConfigDict): config of optimizer
+        log_cfg (ConfigDict): configuration for saving log and checkpoint
+        actor (nn.Module): actor model to select actions
+        actor_target (nn.Module): target actor model to select actions
+        critic (nn.Module): critic model to predict state values
+        critic_target (nn.Module): target critic model to predict state values
+        actor_optim (Optimizer): optimizer for training actor
+        critic_optim (Optimizer): optimizer for training critic
 
     """
 
     def __init__(
-        self, args: argparse.Namespace, hyper_params: ConfigDict, device: torch.device
+        self,
+        args: argparse.Namespace,
+        hyper_params: ConfigDict,
+        log_cfg: ConfigDict,
+        head_cfg: ConfigDict,
+        backbone_cfg: ConfigDict,
+        optim_cfg: ConfigDict,
+        device: torch.device,
     ):
-        Learner.__init__(self, args, hyper_params, device)
+        Learner.__init__(self, args, hyper_params, log_cfg, device)
+
+        self.head_cfg = head_cfg
+        self.backbone_cfg = backbone_cfg
+        self.optim_cfg = optim_cfg
+
+        self._init_network()
+
+    def _init_network(self):
+        """Initialize networks and optimizers."""
+        # create actor
+        self.actor = Brain(self.backbone_cfg.actor, self.head_cfg.actor).to(self.device)
+        self.actor_target = Brain(self.backbone_cfg.actor, self.head_cfg.actor).to(
+            self.device
+        )
+        self.actor_target.load_state_dict(self.actor.state_dict())
+
+        # create critic
+        self.critic = Brain(self.backbone_cfg.critic, self.head_cfg.critic).to(
+            self.device
+        )
+        self.critic_target = Brain(self.backbone_cfg.critic, self.head_cfg.critic).to(
+            self.device
+        )
+        self.critic_target.load_state_dict(self.critic.state_dict())
+
+        # create optimizer
+        self.actor_optim = optim.Adam(
+            self.actor.parameters(),
+            lr=self.optim_cfg.lr_actor,
+            weight_decay=self.optim_cfg.weight_decay,
+        )
+
+        self.critic_optim = optim.Adam(
+            self.critic.parameters(),
+            lr=self.optim_cfg.lr_critic,
+            weight_decay=self.optim_cfg.weight_decay,
+        )
+
+        # load the optimizer and model parameters
+        if self.args.load_from is not None:
+            self.load_params(self.args.load_from)
 
     def update_model(
-        self,
-        networks: Tuple[Brain, ...],
-        optimizer: Union[optim.Optimizer, Tuple[optim.Optimizer, ...]],
-        experience: TensorTuple,
-    ) -> TensorTuple:
+        self, experience: Tuple[torch.Tensor, ...]
+    ) -> Tuple[torch.Tensor, ...]:
         """Update ddpg actor and critic networks"""
-        actor, actor_target, critic, critic_target = networks
-        actor_optim, critic_optim = optimizer
         states, actions, rewards, next_states, dones = experience
 
         # G_t   = r + gamma * v(s_{t+1})  if state != Terminal
         #       = r                       otherwise
         masks = 1 - dones
-        next_actions = actor_target(next_states)
-        next_values = critic_target(torch.cat((next_states, next_actions), dim=-1))
+        next_actions = self.actor_target(next_states)
+        next_values = self.critic_target(torch.cat((next_states, next_actions), dim=-1))
         curr_returns = rewards + self.hyper_params.gamma * next_values * masks
         curr_returns = curr_returns.to(self.device)
 
@@ -49,23 +100,48 @@ class DDPGLearner(Learner):
         gradient_clip_ac = self.hyper_params.gradient_clip_ac
         gradient_clip_cr = self.hyper_params.gradient_clip_cr
 
-        values = critic(torch.cat((states, actions), dim=-1))
+        values = self.critic(torch.cat((states, actions), dim=-1))
         critic_loss = F.mse_loss(values, curr_returns)
-        critic_optim.zero_grad()
+        self.critic_optim.zero_grad()
         critic_loss.backward()
-        clip_grad_norm_(critic.parameters(), gradient_clip_cr)
-        critic_optim.step()
+        clip_grad_norm_(self.critic.parameters(), gradient_clip_cr)
+        self.critic_optim.step()
 
         # train actor
-        actions = actor(states)
-        actor_loss = -critic(torch.cat((states, actions), dim=-1)).mean()
-        actor_optim.zero_grad()
+        actions = self.actor(states)
+        actor_loss = -self.critic(torch.cat((states, actions), dim=-1)).mean()
+        self.actor_optim.zero_grad()
         actor_loss.backward()
-        clip_grad_norm_(actor.parameters(), gradient_clip_ac)
-        actor_optim.step()
+        clip_grad_norm_(self.actor.parameters(), gradient_clip_ac)
+        self.actor_optim.step()
 
         # update target networks
-        common_utils.soft_update(actor, actor_target, self.hyper_params.tau)
-        common_utils.soft_update(critic, critic_target, self.hyper_params.tau)
+        common_utils.soft_update(self.actor, self.actor_target, self.hyper_params.tau)
+        common_utils.soft_update(self.critic, self.critic_target, self.hyper_params.tau)
 
         return actor_loss.item(), critic_loss.item()
+
+    def save_params(self, n_episode: int):
+        """Save model and optimizer parameters."""
+        params = {
+            "actor_state_dict": self.actor.state_dict(),
+            "actor_target_state_dict": self.actor_target.state_dict(),
+            "critic_state_dict": self.critic.state_dict(),
+            "critic_target_state_dict": self.critic_target.state_dict(),
+            "actor_optim_state_dict": self.actor_optim.state_dict(),
+            "critic_optim_state_dict": self.critic_optim.state_dict(),
+        }
+        Learner._save_params(self, params, n_episode)
+
+    def load_params(self, path: str):
+        """Load model and optimizer parameters."""
+        Learner.load_params(self, path)
+
+        params = torch.load(path)
+        self.actor.load_state_dict(params["actor_state_dict"])
+        self.actor_target.load_state_dict(params["actor_target_state_dict"])
+        self.critic.load_state_dict(params["critic_state_dict"])
+        self.critic_target.load_state_dict(params["critic_target_state_dict"])
+        self.actor_optim.load_state_dict(params["actor_optim_state_dict"])
+        self.critic_optim.load_state_dict(params["critic_optim_state_dict"])
+        print("[INFO] loaded the model and optimizer from", path)
