+++ rl_algorithms/dqn/learner.py
@@ -1,5 +1,5 @@
 import argparse
-from typing import Callable, Tuple, Union
+from typing import Tuple, Union
 
 import numpy as np
 import torch
@@ -9,6 +9,7 @@ import torch.optim as optim
 from rl_algorithms.common.abstract.learner import Learner, TensorTuple
 import rl_algorithms.common.helper_functions as common_utils
 from rl_algorithms.common.networks.brain import Brain
+from rl_algorithms.registry import build_loss
 from rl_algorithms.utils.config import ConfigDict
 
 
@@ -18,6 +19,10 @@ class DQNLearner(Learner):
     Attributes:
         args (argparse.Namespace): arguments including hyperparameters and training settings
         hyper_params (ConfigDict): hyper-parameters
+        log_cfg (ConfigDict): configuration for saving log and checkpoint
+        dqn (nn.Module): dqn model to predict state Q values
+        dqn_target (nn.Module): target dqn model to predict state Q values
+        dqn_optim (Optimizer): optimizer for training dqn
 
     """
 
@@ -25,24 +30,46 @@ class DQNLearner(Learner):
         self,
         args: argparse.Namespace,
         hyper_params: ConfigDict,
+        log_cfg: ConfigDict,
         head_cfg: ConfigDict,
-        loss_fn: Callable,
+        backbone_cfg: ConfigDict,
+        optim_cfg: ConfigDict,
         device: torch.device,
     ):
-        Learner.__init__(self, args, hyper_params, device)
+        Learner.__init__(self, args, hyper_params, log_cfg, device)
+
         self.head_cfg = head_cfg
-        self.loss_fn = loss_fn
+        self.backbone_cfg = backbone_cfg
+        self.optim_cfg = optim_cfg
         self.use_n_step = hyper_params.n_step > 1
 
+        self._init_network()
+
+    # pylint: disable=attribute-defined-outside-init
+    def _init_network(self):
+        """Initialize networks and optimizers."""
+        self.dqn = Brain(self.backbone_cfg, self.head_cfg).to(self.device)
+        self.dqn_target = Brain(self.backbone_cfg, self.head_cfg).to(self.device)
+        self.loss_fn = build_loss(self.hyper_params.loss_type)
+
+        self.dqn_target.load_state_dict(self.dqn.state_dict())
+
+        # create optimizer
+        self.dqn_optim = optim.Adam(
+            self.dqn.parameters(),
+            lr=self.optim_cfg.lr_dqn,
+            weight_decay=self.optim_cfg.weight_decay,
+            eps=self.optim_cfg.adam_eps,
+        )
+
+        # load the optimizer and model parameters
+        if self.args.load_from is not None:
+            self.load_params(self.args.load_from)
+
     def update_model(
-        self,
-        networks: Tuple[Brain, ...],
-        optimizer: Union[optim.Optimizer, Tuple[optim.Optimizer, ...]],
-        experience: Union[TensorTuple, Tuple[TensorTuple]],
+        self, experience: Union[TensorTuple, Tuple[TensorTuple]]
     ) -> Tuple[torch.Tensor, torch.Tensor, list, np.ndarray]:  # type: ignore
         """Update dqn and dqn target"""
-        dqn, dqn_target = networks
-        dqn_optim = optimizer
 
         if self.use_n_step:
             experience_1, experience_n = experience
@@ -50,16 +77,20 @@ class DQNLearner(Learner):
             experience_1 = experience
 
         weights, indices = experience_1[-3:-1]
+
         gamma = self.hyper_params.gamma
+
         dq_loss_element_wise, q_values = self.loss_fn(
-            dqn, dqn_target, experience_1, gamma, self.head_cfg
+            self.dqn, self.dqn_target, experience_1, gamma, self.head_cfg
         )
         dq_loss = torch.mean(dq_loss_element_wise * weights)
 
+        # n step loss
         if self.use_n_step:
             gamma = self.hyper_params.gamma ** self.hyper_params.n_step
+
             dq_loss_n_element_wise, q_values_n = self.loss_fn(
-                dqn, dqn_target, experience_n, gamma, self.head_cfg
+                self.dqn, self.dqn_target, experience_n, gamma, self.head_cfg
             )
 
             # to update loss and priorities
@@ -73,20 +104,46 @@ class DQNLearner(Learner):
         # total loss
         loss = dq_loss + q_regular
 
-        dqn_optim.zero_grad()
+        self.dqn_optim.zero_grad()
         loss.backward()
-        clip_grad_norm_(dqn.parameters(), self.hyper_params.gradient_clip)
-        dqn_optim.step()
+        clip_grad_norm_(self.dqn.parameters(), self.hyper_params.gradient_clip)
+        self.dqn_optim.step()
 
         # update target networks
-        common_utils.soft_update(dqn, dqn_target, self.hyper_params.tau)
+        common_utils.soft_update(self.dqn, self.dqn_target, self.hyper_params.tau)
 
         # update priorities in PER
         loss_for_prior = dq_loss_element_wise.detach().cpu().numpy()
         new_priorities = loss_for_prior + self.hyper_params.per_eps
 
         if self.head_cfg.configs.use_noisy_net:
-            dqn.head.reset_noise()
-            dqn_target.head.reset_noise()
+            self.dqn.head.reset_noise()
+            self.dqn_target.head.reset_noise()
 
-        return loss.item(), q_values.mean().item(), indices, new_priorities
+        return (
+            loss.item(),
+            q_values.mean().item(),
+            indices.long().cpu().numpy(),
+            new_priorities,
+        )
+
+    def save_params(self, n_episode: int):
+        """Save model and optimizer parameters."""
+        params = {
+            "dqn_state_dict": self.dqn.state_dict(),
+            "dqn_target_state_dict": self.dqn_target.state_dict(),
+            "dqn_optim_state_dict": self.dqn_optim.state_dict(),
+        }
+
+        Learner._save_params(self, params, n_episode)
+
+    # pylint: disable=attribute-defined-outside-init
+    def load_params(self, path: str):
+        """Load model and optimizer parameters."""
+        Learner.load_params(self, path)
+
+        params = torch.load(path)
+        self.dqn.load_state_dict(params["dqn_state_dict"])
+        self.dqn_target.load_state_dict(params["dqn_target_state_dict"])
+        self.dqn_optim.load_state_dict(params["dqn_optim_state_dict"])
+        print("[INFO] loaded the model and optimizer from", path)
