+++ rl_algorithms/ppo/learner.py
@@ -1,5 +1,4 @@
 import argparse
-from typing import Tuple, Union
 
 import torch
 from torch.nn.utils import clip_grad_norm_
@@ -17,6 +16,11 @@ class PPOLearner(Learner):
     Attributes:
         args (argparse.Namespace): arguments including hyperparameters and training settings
         hyper_params (ConfigDict): hyper-parameters
+        log_cfg (ConfigDict): configuration for saving log and checkpoint
+        actor (nn.Module): actor model to select actions
+        critic (nn.Module): critic model to predict state values
+        actor_optim (Optimizer): optimizer for training actor
+        critic_optim (Optimizer): optimizer for training critic
 
     """
 
@@ -24,27 +28,52 @@ class PPOLearner(Learner):
         self,
         args: argparse.Namespace,
         hyper_params: ConfigDict,
+        log_cfg: ConfigDict,
+        head_cfg: ConfigDict,
+        backbone_cfg: ConfigDict,
+        optim_cfg: ConfigDict,
         device: torch.device,
         is_discrete: bool,
     ):
-        Learner.__init__(self, args, hyper_params, device)
+        Learner.__init__(self, args, hyper_params, log_cfg, device)
 
+        self.head_cfg = head_cfg
+        self.backbone_cfg = backbone_cfg
+        self.optim_cfg = optim_cfg
         self.is_discrete = is_discrete
 
-    def update_model(
-        self,
-        networks: Tuple[Brain, ...],
-        optimizer: Union[optim.Optimizer, Tuple[optim.Optimizer, ...]],
-        experience: TensorTuple,
-        epsilon: float,
-    ) -> TensorTuple:
-        """Update PPO actor and critic networks"""
-        actor, critic = networks
-        actor_optim, critic_optim = optimizer
+        self._init_network()
 
+    def _init_network(self):
+        """Initialize networks and optimizers."""
+        # create actor
+        self.actor = Brain(self.backbone_cfg.actor, self.head_cfg.actor).to(self.device)
+        self.critic = Brain(self.backbone_cfg.critic, self.head_cfg.critic).to(
+            self.device
+        )
+
+        # create optimizer
+        self.actor_optim = optim.Adam(
+            self.actor.parameters(),
+            lr=self.optim_cfg.lr_actor,
+            weight_decay=self.optim_cfg.weight_decay,
+        )
+
+        self.critic_optim = optim.Adam(
+            self.critic.parameters(),
+            lr=self.optim_cfg.lr_critic,
+            weight_decay=self.optim_cfg.weight_decay,
+        )
+
+        # load model parameters
+        if self.args.load_from is not None:
+            self.load_params(self.args.load_from)
+
+    def update_model(self, experience: TensorTuple, epsilon: float) -> TensorTuple:
+        """Update PPO actor and critic networks"""
         states, actions, rewards, values, log_probs, next_state, masks = experience
         next_state = torch.FloatTensor(next_state).to(self.device)
-        next_value = critic(next_state)
+        next_value = self.critic(next_state)
 
         returns = ppo_utils.compute_gae(
             next_value,
@@ -82,7 +111,7 @@ class PPOLearner(Learner):
             advantages,
         ):
             # calculate ratios
-            _, dist = actor(state)
+            _, dist = self.actor(state)
             log_prob = dist.log_prob(action)
             ratio = (log_prob - old_log_prob).exp()
 
@@ -92,7 +121,7 @@ class PPOLearner(Learner):
             actor_loss = -torch.min(surr_loss, clipped_surr_loss).mean()
 
             # critic_loss
-            value = critic(state)
+            value = self.critic(state)
             if self.hyper_params.use_clipped_value_loss:
                 value_pred_clipped = old_value + torch.clamp(
                     (value - old_value), -epsilon, epsilon
@@ -116,16 +145,16 @@ class PPOLearner(Learner):
             gradient_clip_ac = self.hyper_params.gradient_clip_ac
             gradient_clip_cr = self.hyper_params.gradient_clip_cr
 
-            critic_optim.zero_grad()
+            self.critic_optim.zero_grad()
             total_loss.backward(retain_graph=True)
-            clip_grad_norm_(critic.parameters(), gradient_clip_ac)
-            critic_optim.step()
+            clip_grad_norm_(self.critic.parameters(), gradient_clip_ac)
+            self.critic_optim.step()
 
             # train actor
-            actor_optim.zero_grad()
+            self.actor_optim.zero_grad()
             total_loss.backward()
-            clip_grad_norm_(critic.parameters(), gradient_clip_cr)
-            actor_optim.step()
+            clip_grad_norm_(self.actor.parameters(), gradient_clip_cr)
+            self.actor_optim.step()
 
             actor_losses.append(actor_loss.item())
             critic_losses.append(critic_loss.item())
@@ -136,3 +165,24 @@ class PPOLearner(Learner):
         total_loss = sum(total_losses) / len(total_losses)
 
         return actor_loss, critic_loss, total_loss
+
+    def save_params(self, n_episode: int):
+        """Save model and optimizer parameters."""
+        params = {
+            "actor_state_dict": self.actor.state_dict(),
+            "critic_state_dict": self.critic.state_dict(),
+            "actor_optim_state_dict": self.actor_optim.state_dict(),
+            "critic_optim_state_dict": self.critic_optim.state_dict(),
+        }
+        Learner._save_params(self, params, n_episode)
+
+    def load_params(self, path: str):
+        """Load model and optimizer parameters."""
+        Learner.load_params(self, path)
+
+        params = torch.load(path)
+        self.actor.load_state_dict(params["actor_state_dict"])
+        self.critic.load_state_dict(params["critic_state_dict"])
+        self.actor_optim.load_state_dict(params["actor_optim_state_dict"])
+        self.critic_optim.load_state_dict(params["critic_optim_state_dict"])
+        print("[INFO] loaded the model and optimizer from", path)
