+++ rl_algorithms/sac/agent.py
@@ -14,12 +14,11 @@ from typing import Tuple
 import gym
 import numpy as np
 import torch
-import torch.optim as optim
 import wandb
 
 from rl_algorithms.common.abstract.agent import Agent
 from rl_algorithms.common.buffer.replay_buffer import ReplayBuffer
-from rl_algorithms.common.networks.brain import Brain
+from rl_algorithms.common.helper_functions import numpy2floattensor
 from rl_algorithms.registry import AGENTS
 from rl_algorithms.sac.learner import SACLearner
 from rl_algorithms.utils.config import ConfigDict
@@ -40,14 +39,6 @@ class SACAgent(Agent):
         state_dim (int): state size of env
         action_dim (int): action size of env
         memory (ReplayBuffer): replay memory
-        actor (nn.Module): actor model to select actions
-        actor_optim (Optimizer): optimizer for training actor
-        critic_1 (nn.Module): critic model to predict state values
-        critic_2 (nn.Module): critic model to predict state values
-        critic_target1 (nn.Module): target critic model to predict state values
-        critic_target2 (nn.Module): target critic model to predict state values
-        critic_optim1 (Optimizer): optimizer for training critic_1
-        critic_optim2 (Optimizer): optimizer for training critic_2
         curr_state (np.ndarray): temporary storage of the current state
         total_step (int): total step numbers
         episode_step (int): step number of the current episode
@@ -88,19 +79,16 @@ class SACAgent(Agent):
 
         self.state_dim = self.env.observation_space.shape
         self.action_dim = self.env.action_space.shape[0]
-        self.hyper_params["state_dim"] = self.state_dim
-        self.hyper_params["action_dim"] = self.action_dim
-
-        # target entropy
-        target_entropy = -np.prod((self.action_dim,)).item()  # heuristic
-        # automatic entropy tuning
-        if hyper_params.auto_entropy_tuning:
-            self.target_entropy = target_entropy
-            self.log_alpha = torch.zeros(1, requires_grad=True, device=device)
-            self.alpha_optim = optim.Adam([self.log_alpha], lr=optim_cfg.lr_entropy)
+        self.head_cfg["action_dim"] = self.action_dim
+        self.head_cfg.actor.configs.state_size = (
+            self.head_cfg.critic_vf.configs.state_size
+        ) = self.state_dim
+        self.head_cfg.critic_qf.configs.state_size = (
+            self.state_dim[0] + self.action_dim,
+        )
+        self.head_cfg.actor.configs.output_size = self.action_dim
 
         self._initialize()
-        self._init_network()
 
     # pylint: disable=attribute-defined-outside-init
     def _initialize(self):
@@ -111,63 +99,15 @@ class SACAgent(Agent):
                 self.hyper_params.buffer_size, self.hyper_params.batch_size
             )
 
-    # pylint: disable=attribute-defined-outside-init
-    def _init_network(self):
-        """Initialize networks and optimizers."""
-
-        self.head_cfg.actor.configs.state_size = (
-            self.head_cfg.critic_vf.configs.state_size
-        ) = self.state_dim
-        self.head_cfg.critic_qf.configs.state_size = (
-            self.state_dim[0] + self.action_dim,
+        self.learner = SACLearner(
+            self.args,
+            self.hyper_params,
+            self.log_cfg,
+            self.head_cfg,
+            self.backbone_cfg,
+            self.optim_cfg,
+            device,
         )
-        self.head_cfg.actor.configs.output_size = self.action_dim
-
-        # create actor
-        self.actor = Brain(self.backbone_cfg.actor, self.head_cfg.actor).to(device)
-
-        # create v_critic
-        self.vf = Brain(self.backbone_cfg.critic_vf, self.head_cfg.critic_vf).to(device)
-        self.vf_target = Brain(self.backbone_cfg.critic_vf, self.head_cfg.critic_vf).to(
-            device
-        )
-        self.vf_target.load_state_dict(self.vf.state_dict())
-
-        # create q_critic
-        self.qf_1 = Brain(self.backbone_cfg.critic_qf, self.head_cfg.critic_qf).to(
-            device
-        )
-        self.qf_2 = Brain(self.backbone_cfg.critic_qf, self.head_cfg.critic_qf).to(
-            device
-        )
-
-        # create optimizers
-        self.actor_optim = optim.Adam(
-            self.actor.parameters(),
-            lr=self.optim_cfg.lr_actor,
-            weight_decay=self.optim_cfg.weight_decay,
-        )
-        self.vf_optim = optim.Adam(
-            self.vf.parameters(),
-            lr=self.optim_cfg.lr_vf,
-            weight_decay=self.optim_cfg.weight_decay,
-        )
-        self.qf_1_optim = optim.Adam(
-            self.qf_1.parameters(),
-            lr=self.optim_cfg.lr_qf1,
-            weight_decay=self.optim_cfg.weight_decay,
-        )
-        self.qf_2_optim = optim.Adam(
-            self.qf_2.parameters(),
-            lr=self.optim_cfg.lr_qf2,
-            weight_decay=self.optim_cfg.weight_decay,
-        )
-
-        # load the optimizer and model parameters
-        if self.args.load_from is not None:
-            self.load_params(self.args.load_from)
-
-        self.learner = SACLearner(self.args, self.hyper_params, device)
 
     def select_action(self, state: np.ndarray) -> np.ndarray:
         """Select an action from the input space."""
@@ -182,9 +122,9 @@ class SACAgent(Agent):
             return np.array(self.env.action_space.sample())
 
         if self.args.test:
-            _, _, _, selected_action, _ = self.actor(state)
+            _, _, _, selected_action, _ = self.learner.actor(state)
         else:
-            selected_action, _, _, _, _ = self.actor(state)
+            selected_action, _, _, _, _ = self.learner.actor(state)
 
         return selected_action.detach().cpu().numpy()
 
@@ -211,134 +151,6 @@ class SACAgent(Agent):
     def _add_transition_to_memory(self, transition: Tuple[np.ndarray, ...]):
         """Add 1 step and n step transitions to memory."""
         self.memory.add(transition)
-
-    def update_model(self) -> Tuple[torch.Tensor, ...]:
-        """Train the model after each episode."""
-        self.update_step += 1
-
-        experiences = self.memory.sample()
-        experiences = self.numpy2floattensor(experiences)
-        states, actions, rewards, next_states, dones = experiences
-        new_actions, log_prob, pre_tanh_value, mu, std = self.actor(states)
-
-        # train alpha
-        if self.hyper_params.auto_entropy_tuning:
-            alpha_loss = (
-                -self.log_alpha * (log_prob + self.target_entropy).detach()
-            ).mean()
-
-            self.alpha_optim.zero_grad()
-            alpha_loss.backward()
-            self.alpha_optim.step()
-
-            alpha = self.log_alpha.exp()
-        else:
-            alpha_loss = torch.zeros(1)
-            alpha = self.hyper_params.w_entropy
-
-        # Q function loss
-        masks = 1 - dones
-        states_actions = torch.cat((states, actions), dim=-1)
-        q_1_pred = self.qf_1(states_actions)
-        q_2_pred = self.qf_2(states_actions)
-        v_target = self.vf_target(next_states)
-        q_target = rewards + self.hyper_params.gamma * v_target * masks
-        qf_1_loss = F.mse_loss(q_1_pred, q_target.detach())
-        qf_2_loss = F.mse_loss(q_2_pred, q_target.detach())
-
-        # V function loss
-        states_actions = torch.cat((states, new_actions), dim=-1)
-        v_pred = self.vf(states)
-        q_pred = torch.min(self.qf_1(states_actions), self.qf_2(states_actions))
-        v_target = q_pred - alpha * log_prob
-        vf_loss = F.mse_loss(v_pred, v_target.detach())
-
-        # train Q functions
-        self.qf_1_optim.zero_grad()
-        qf_1_loss.backward()
-        self.qf_1_optim.step()
-
-        self.qf_2_optim.zero_grad()
-        qf_2_loss.backward()
-        self.qf_2_optim.step()
-
-        # train V function
-        self.vf_optim.zero_grad()
-        vf_loss.backward()
-        self.vf_optim.step()
-
-        if self.update_step % self.hyper_params.policy_update_freq == 0:
-            # actor loss
-            advantage = q_pred - v_pred.detach()
-            actor_loss = (alpha * log_prob - advantage).mean()
-
-            # regularization
-            mean_reg = self.hyper_params.w_mean_reg * mu.pow(2).mean()
-            std_reg = self.hyper_params.w_std_reg * std.pow(2).mean()
-            pre_activation_reg = self.hyper_params.w_pre_activation_reg * (
-                pre_tanh_value.pow(2).sum(dim=-1).mean()
-            )
-            actor_reg = mean_reg + std_reg + pre_activation_reg
-
-            # actor loss + regularization
-            actor_loss += actor_reg
-
-            # train actor
-            self.actor_optim.zero_grad()
-            actor_loss.backward()
-            self.actor_optim.step()
-
-            # update target networks
-            common_utils.soft_update(self.vf, self.vf_target, self.hyper_params.tau)
-        else:
-            actor_loss = torch.zeros(1)
-
-        return (
-            actor_loss.item(),
-            qf_1_loss.item(),
-            qf_2_loss.item(),
-            vf_loss.item(),
-            alpha_loss.item(),
-        )
-
-    def load_params(self, path: str):
-        """Load model and optimizer parameters."""
-        Agent.load_params(self, path)
-
-        params = torch.load(path)
-        self.actor.load_state_dict(params["actor"])
-        self.qf_1.load_state_dict(params["qf_1"])
-        self.qf_2.load_state_dict(params["qf_2"])
-        self.vf.load_state_dict(params["vf"])
-        self.vf_target.load_state_dict(params["vf_target"])
-        self.actor_optim.load_state_dict(params["actor_optim"])
-        self.qf_1_optim.load_state_dict(params["qf_1_optim"])
-        self.qf_2_optim.load_state_dict(params["qf_2_optim"])
-        self.vf_optim.load_state_dict(params["vf_optim"])
-
-        if self.hyper_params.auto_entropy_tuning:
-            self.alpha_optim.load_state_dict(params["alpha_optim"])
-
-        print("[INFO] loaded the model and optimizer from", path)
-
-    def save_params(self, n_episode: int):
-        """Save model and optimizer parameters."""
-        params = {
-            "actor": self.actor.state_dict(),
-            "qf_1": self.qf_1.state_dict(),
-            "qf_2": self.qf_2.state_dict(),
-            "vf": self.vf.state_dict(),
-            "vf_target": self.vf_target.state_dict(),
-            "actor_optim": self.actor_optim.state_dict(),
-            "qf_1_optim": self.qf_1_optim.state_dict(),
-            "qf_2_optim": self.qf_2_optim.state_dict(),
-            "vf_optim": self.vf_optim.state_dict(),
-        }
-
-        if self.hyper_params.auto_entropy_tuning:
-            params["alpha_optim"] = self.alpha_optim.state_dict()
-
-        Agent._save_params(self, params, n_episode)
 
     def write_log(self, log_value: tuple):
         """Write log about loss and score"""
@@ -418,16 +230,8 @@ class SACAgent(Agent):
                 if len(self.memory) >= self.hyper_params.batch_size:
                     for _ in range(self.hyper_params.multiple_update):
                         experience = self.memory.sample()
-                        loss = self.learner.update_model(
-                            (self.actor, self.vf, self.vf_target, self.qf_1, self.qf_2),
-                            (
-                                self.actor_optim,
-                                self.vf_optim,
-                                self.qf_1_optim,
-                                self.qf_2_optim,
-                            ),
-                            experience,
-                        )
+                        experience = numpy2floattensor(experience)
+                        loss = self.learner.update_model(experience)
                         loss_episode.append(loss)  # for logging
 
             t_end = time.time()
@@ -446,10 +250,10 @@ class SACAgent(Agent):
                 self.write_log(log_value)
 
             if self.i_episode % self.args.save_period == 0:
-                self.save_params(self.i_episode)
+                self.learner.save_params(self.i_episode)
                 self.interim_test()
 
         # termination
         self.env.close()
-        self.save_params(self.i_episode)
+        self.learner.save_params(self.i_episode)
         self.interim_test()
